Following the readme from a fresh `git clone`:
When `npm install`ing, npm warns of some missing peer deps. Webpack
compilation fails without them because of these deps:

+ bootstrap-vue:
  - jquery [missing]
  - popper.js [missing]

Additionally, `node-sass` (also a dependecy of bootstrap) was in a weird state of missing a `vendor` directory.
This could be unrelated but I thought it worth mentioning.
After updating `package.json` and `npm i` again, it has the directory
and everything works.